Extract inspect query runner partial
